[BUGFIX] Centering images with text is not working in some cases
Currently css_styled_content is using the constant MaxW in CSS to define the width of the image wrap. When using columns with different widths, this value is not reflecting the width of the column, which can break the centering.

This patch fixes this issue by using a different centering method.

Fixes: #35814

Change-Id: I2d1532b986054b8ae250d8ce6cab5b67c61a6c5d
Releases: 6.0, 4.7
Reviewed-on: http://review.typo3.org/10256
Reviewed-by: Berit Jensen
Tested-by: Berit Jensen
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter