[FEATURE] Registers for title & alt text fields in file links
The file link element offers a description, title text and alternative
label text for each file.
Currently, it is only possible to access the description via a register.
It would be desirable to also provide registers for the alt and title
text fields, e.g., to only link the description and provide more
in-depth info in the other fields.

Change-Id: I7860a769e957c83691e48090ea1bbc94c9ebf9df
Resolves: #44182
Releases: 6.2
Reviewed-on: https://review.typo3.org/17705
Reviewed-by: Alexander Opitz
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer