[BUGFIX] header element is rendered in lib.stdheader in all HTML5 cases
According to the HTML5 specs the header element should be used for
grouping elements. Currently the header element is used if there is any
content in lib.stdheader, for instance a single hx.

This patch fixes this by wrapping lib.stdheader with the header element
only when the date is added to the header. Otherwise it will use the
default div wrapping.

Change-Id: I3e1335e1b1d64962a8488b8177d72eb1126e89dd
Fixes: #47575
Releases: 6.2
Reviewed-on: https://review.typo3.org/20149
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny
Reviewed-by: Kai Ole Hartwig
Tested-by: Kai Ole Hartwig
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters