[TASK] tt_content.shortcut depends on layout
CSS Styled Content delivers its TypoScript
and got a cleanup in 4.6 due to removal
of old plugins.
However, the replacement (in the default
TS from 4.6 to 6.2) has superfluous code
and is not backwards-compatible if the
layout field is used differently or not used
at all (e.g. NULL value instead of 0) for
installations that have been upgraded
since 4.0.

The patch simplifies and unifies
the default TypoScript for
tt_content.shortcut.

Releases: 6.2
Change-Id: I3ed49e265329155af683a993476ffbb13226462c
Reviewed-on: https://review.typo3.org/27211
Reviewed-by: Sebastian Dudda
Tested-by: Sebastian Dudda
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack