[!!!][TASK] Move old mailform to compatibility6
Due to the replacement of the better
EXT:form logic, the fallback and default
mailform when EXT:form is not installed,
is moved to compatibility6.

There are certain parts that are migrated
as well and therefore removed from the
main core installation.

 * Various methods in ContentObjectRenderer,
and especially TypoScriptFrontendRenderer
which specifically served the purpose.
 * The FORM Content Object
 * The TCA tt_content mailform configuration
 * The TypoScript from css styled content
 * The newContentElement wizard configuration
 * Related tests to the form-related methods
 * The preview functionality for a mailform in "Web"=>"Page"

The submission class (DataSubmissionController)
has been moved and renamed as well, and now
hooks into the checkDataSubmission of TSFE.

Additionally, the related tests to the methods
have been removed.

Resolves: #64668
Releases: master
Change-Id: Iefbaf11322678f811d5c38d82f7d81be56b1d378
Reviewed-on: http://review.typo3.org/36479
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>