[BUGFIX] Remove inline styles from parsed blockquote tag
When passing blockquote tags trough lib.parseFunc_RTE (this
is commonly done when rendering TS objects in Fluid templates)
their attributes will be replaced with some inline styles.
I don't know what the intention of this behavior is, but it's been
like this at least since 2004. Inline styles should generally be
avoided.
I suggest to simple remove this rule.

Code to reproduce the issue and test the change:
https://gist.github.com/4654874

Resolves: #44879
Releases: master
Change-Id: If8954fa392b906d78a8cc32aba9e5b23d84decd9
Reviewed-on: http://review.typo3.org/17758
Reviewed-by: Steffen Müller <typo3@t3node.com>
Tested-by: Steffen Müller <typo3@t3node.com>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>