[!!!][FEATURE] TypoScript select is now translated by default
The TypoScript .select option which is used for Content Objects
like "CONTENT", has the property "languageField". This allows
to set the field that has the information about the sys_language_uid
value in order to have only records shown that are translated or
on -1 (show in all languages).

In the past, this option had to be set explicitly but is now turned
on by default, but can turned off, if this behavior is not desired.

Besides this patch adds some basic unit tests for the function and
resolves and issue with groupBy which popped up because of the tests.

Resolves: #68191
Releases: master
Change-Id: I36765d9e1e8a8b49fdc05700caa1e2e3f613a06d
Reviewed-on: http://review.typo3.org/41370
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>