[TASK] Unify TypoScript-related helper methods
The method for OptionSplitting in TemplateService
is completely isolated from everything else, and creates
a very strong dependency to TSFE + TemplateService
in various places.

Additionally, the TypoScriptService of Extbase is used
in several places of the core and has nothing to do
with Extbase itself, thus, it can be moved to a central place
where these functions are called.

Moving this code to it's own service class in the core
helps to decouple TSFE, tmpl, Extbase and cObject more.

Resolves: #78650
Releases: master
Change-Id: I712c77c383171e93053caff0f89af8d1247f6d86
Reviewed-on: https://review.typo3.org/50589
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>