[BUGFIX] Allow t3:// syntax to be parsed by lib.parseFunc
When introducing the new t3:// linking syntax for typolink,
we also allowed to use them in the RTE without using <link>
anymore, which is really cool. However, the frontend (= TypoScript)
cannot deal with that yet.

A clean function within parseFunc should "transformLinks" with typolink.
For the time being, a parseFunc.tags.A is introduced to allow
the transformations to be rendered correctly.

It *could* have a side-effect that certain attributes
within parseFunc.tags.A are rendered twice, but can be solved
at a later point by sanitizing the attributes quite simple.

Resolves: #79041
Releases: master
Change-Id: Ib8da2a88b67a633fc89a5c1ddee90c8e14f5b27b
Reviewed-on: https://review.typo3.org/51000
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>